pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
# dependencies
loguru = "==0.7.2"
python-dotenv = "1.0.1"
kafka-python = "2.0.2"
pymemcache = "4.0.0"
msgspec = "0.18.6"
yarl = "1.9.6"
bgm-tv-wiki = "^0.0.12"
sqlalchemy = "^2.0.34"
pymysql = "^1.1.1"

[tool.poetry.group.dev.dependencies]
# linter and formatter
pre-commit = "==3.8.0"
mypy = "==1.11.1"
watchgod = "0.8.2"
pytest = "^8.3.2"
pytest-github-actions-annotate-failures = "^0.2.0"

[tool.poetry-plugin-bump]
commit_msg = 'bump: v{version}'

[tool.coverage.run]
concurrency = ['thread', 'greenlet']

[tool.coverage.report]
exclude_lines = [
    'pragma: no cover',
    'if TYPE_CHECKING:',
    'if __name__ == "__main__":',
    'raise NotImplementedError'
]

[tool.pytest.ini_options]
addopts = '-rav'

[tool.mypy]
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
show_error_codes = true
pretty = true
plugins = []

[tool.black]
target_version = ['py310']

[tool.ruff]
target-version = "py310"
extend-exclude = [".venv", "api"]

[tool.ruff.lint]
extend-select = ['I']

[tool.ruff.lint.isort]
length-sort = true
lines-after-imports = 2
